requirements.txt
ansible==2.1.2.0
apt-xapian-index==0.47
apturl==0.5.2
astroid==1.4.4
chardet==2.3.0
command-not-found==0.3
defer==1.0.6
Django==1.11.1
django-cors-headers==2.1.0
django-filter==1.0.4
djangorestframework==3.6.3
language-selector==0.1
lazy-object-proxy==1.2.1
logilab-common==1.1.0
Markdown==2.6.8
pexpect==4.0.1
Pillow==3.1.2
ptyprocess==0.5
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
pylint==1.5.2
python-apt==1.1.0b1
python-debian==0.1.27
python-memcached==1.58
python-systemd==231
pytz==2017.2
PyYAML==3.12
reportlab==3.3.0
requests==2.9.1
simplejson==3.10.0
six==1.10.0
slumber==0.7.1
sqlparse==0.1.18
ssh-import-id==5.5
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.13.1
usb-creator==0.3.0
virtualenv==15.0.1
wrapt==1.8.0
xkit==0.0.0
